Image issue solved

+++ src/utils/cloudinary.js
@@ -1,22 +1,29 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
-cloudinary.config({
+const uploadOnCloudinary=async(localFilePath)=>{
+    
+    cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
     api_key: process.env.CLOUDINARY_API_KEY, 
     api_secret: process.env.CLOUDINARY_API_SECRET
 })
 
-const uploadOnCloudinary=async(localFilePath)=>{
 try {
     if(!localFilePath)return null;
     const response=await cloudinary.uploader.upload(localFilePath,{resource_type:"auto"});
-    fs.unlinkSync(localFilePath);
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath); 
+    }
     return response;
 } catch (error) {
-    fs.unlinkSync(localFilePath);
+     console.error("Cloudinary upload error:", error.message); 
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);  
+    }
     return null;
 }
 }
 
 export {uploadOnCloudinary};
+
